PBC function working.
The PBC function now seems to be working. It was previously wrong because I had been applying boundary conditions to the particles' positions at the *previous* timestep instead of the current timestep. I've since fixed this and added an assert which is false if the particle strays outside of the computational domain.