Added scripts 'clustering_fns.py', 'dynamic_objects.py' and 'vic.py'
The first is a collection of functions used to perform k-means clustering on kinematic data (ie position and velocity of particles). The latter (ie vic.py) is a script which implements a simple Vicsek model of particle motion. 'dynamic_objects.py' is a script which runs the Vicsek simulation and identifies clusters at each timestep.