Modified 'get_edges_vertices' function, 'generate_mol_file' function and deleted 'plot_results' function.
(1) I modified the function 'get_edges_vertices' to accept a single cluster as an argument. I then extract the x and y positions from the cluster dataframe with the line 'cluster_positions = cluster[ ['x-position', 'y-position'] ]'. (2) I modified the function 'generate_mole_file' to accept an argument 'clusters' which is a list whose elements are the dataframes of all clusters. I then iterate over all the clusters in the list and produce mol files for each of these clusters, calling the function 'get_edges_vertices' in each iteration of the loop. (3) In the loop of the main function, I no longer call the function 'get_edges_vertices'. Instead, I call the function 'generate_mol_file' which in turn calls the aforementioned function. (4) I deleted the function 'plot_results'.