Created class specifically for clusters.
New class takes a Pandas dataframe 'cluster_df' as an __init__ argument. The class then creates two attributes, 'clusterDataFrame' and 'numParticles'. The class has methods 'computeCentroid' and 'getAssemblyIndex'. Upon execution of these two methods, the centroid and assembly index of the cluster become attributes of the instance. In the time-stepping part of the main function, I first create an empty dictionary 'cluster_dict' which is then filled after we call 'clus.makeClustersDbscan'. The dictionary 'cluster_dict' has integers as keys (the key is simply the order in which the class variable was initialized) and the value is an object of type cluster. As a final note, I changed the functions 'cluster_histogram' and 'visualize_clusters' to take 'cluster_dict's as arguments instead of 'clusters'.