Modified 'computeEdgesVertices' function to remove warning.
Upon calling the aforementioned function, a warning was generated complaining about valid feature names. This was due to the fact that the data structure passed to the 'kneighbors_graph' function was a pandas dataframe - which associates names with each column in the dataframe. To fix this problem, one can simply extract the desired columns in the dataframe as numpy arrays, appending the '.values' method to the dataframe indexing. In this case, 'clusterDataFrame[ ['x-position', 'y-position'] ]' -> 'clusterDataFrame[ ['x-position', 'y-position'] ].values'.