Created class for dynamic_objects.
The class has as its main attribute a dictionary 'data_dict' which has keys 'CoM' (center of mass - centroid), 'numParticles', 'assemblyIndex', 'area', and 'lifeTime'. The arguments to the class upon initialization are centroidPosition, numParticles, assemblyIndex, and area. All of which come from an object of type cluster. The 'area' refers to the area of the convex hull of all particles in a cluster. An object is only initialized if the centroid of a cluster has moved less than a distance epsilon from the previous timestep. If this condition is met, the cluster is an object that has a lifetime of at least two timesteps. To determine if any cluster C0 at time t_i (i > 1) satisfies this criterion, we loop over all the clusters in the previous timestep and see if the distance between the centroid of C0 and the centroid of any cluster in the previous timestep is less than epsilon. If this condition is met, we initialize an object of dynamic_object type.