Making functionality to classify objects as dead or alive.
By 'alive', I mean a dynamic object that exists in the current time-step. Any object not satisfying this criterion is classified as 'dead'. My current approach is to create a dictionary DoA_dict where the key is the identifier of the object (an integer), and the value is a 0 or 1 - 1 if the object is alive and 0 if dead. As intimated above, keys in the object_dict dictionary are now just integers. Specifically, if an object has a key n, that means the object was the nth distinct object to be identified.