Have added functionality which classifies objects as alive or dead in each timestep
In each timestep, I will have n(t) total objects - both living and dead. At the end of each time step, I create a dictionary 'DoA_dict' which has n(t) entries. The keys of which are the keys in the 'object_dict' dictionary and the values are all initialized to 0. If, in the next time step, a dynamic object is updated, DoA_dict[key] = 1, that is, the object is classified as alive. Then, before re-initializing the DoA_dict dictionary, I loop over all objects in the 'object-dict' dictionary and if any of corresponding values in the 'DoA_dict' dictionary are 0, I modify the object dyn_obj such that dyn_obj.DoA = Dead - that is, I remark that the object is deceased.