Modified the program so that it can be run on both windows and linux machines.
I modified the program so that the main function has a local variable OS such that OS = win or nix which the user must select. Depending on which value of OS is selected, a different executable file will be run when it's time to compute the assembly index of clusters.